ci: remove merge queue
