+++ package.json
@@ -21,10 +21,8 @@
     "check:patches": "node scripts/check/patches.ts",
     "check:site": "pnpm --filter www exec vitest run --config vitest.afdocs.mts --coverage.enabled=false",
     "check:tests": "node scripts/check/tests.ts",
-    "ci:gate": "node scripts/github/queue/main.ts gate",
+    "ci:gate": "node scripts/github/gate.ts",
     "ci:production-acceptance": "node scripts/production/acceptance.ts",
-    "ci:queue": "node scripts/github/queue/main.ts admit",
-    "ci:review": "node scripts/github/queue/main.ts review",
     "clean": "rm -rf node_modules apps/*/node_modules packages/*/node_modules .cache apps/*/.cache packages/*/.cache .pnpm-store apps/*/.turbo packages/*/.turbo apps/*/.ruff_cache apps/*/.pytest_cache apps/*/.venv apps/*/.coverage .turbo .next apps/*/.next apps/*/.react-email packages/*/.react-email .vercel/output apps/*/.vercel/output coverage apps/*/coverage packages/*/coverage && find apps packages -name __pycache__ -type d -prune -exec rm -rf {} +",
     "clean:ts": "find . -name '*.tsbuildinfo' -type f -delete && find . -name 'tsconfig.tsbuildinfo' -type f -delete",
     "deps-check-api": "cd apps/api && depcheck --ignore-dirs=node_modules || true",
@@ -44,7 +42,7 @@
     "start": "turbo run start",
     "test": "pnpm check:patches && pnpm check:tests && pnpm test:scripts && turbo run test",
     "test:coverage": "pnpm check:patches && pnpm check:tests && turbo run test:coverage",
-    "test:scripts": "pnpm typecheck:scripts && pnpm --filter @repo/testing exec vitest run --root ../.. scripts/check/effect.test.ts scripts/dependencies/policy.test.ts scripts/github/cli.test.ts scripts/github/policy.test.ts scripts/github/provenance/bundle.test.ts scripts/github/provenance/verify.test.ts scripts/github/queue/admission.test.ts scripts/github/queue/release.test.ts scripts/github/queue/remote.test.ts scripts/github/queue/replay.test.ts scripts/github/queue/result.test.ts scripts/github/queue/tree.test.ts scripts/github/release.test.ts scripts/vercel/scope.test.ts scripts/effect/source.test.ts scripts/production/acceptance.test.ts",
+    "test:scripts": "pnpm typecheck:scripts && pnpm --filter @repo/testing exec vitest run --root ../.. scripts/check/effect.test.ts scripts/dependencies/policy.test.ts scripts/github/cli.test.ts scripts/github/gate.test.ts scripts/github/policy.test.ts scripts/github/provenance/bundle.test.ts scripts/github/provenance/verify.test.ts scripts/github/release.test.ts scripts/vercel/scope.test.ts scripts/effect/source.test.ts scripts/production/acceptance.test.ts",
     "test:ui": "turbo run test:ui",
     "test:watch": "turbo run test:watch",
     "translate": "turbo run translate",
